v1.1.0 update

+++ src/main/java/com/demo/api/BookInterface.java
@@ -14,5 +14,5 @@ public interface BookInterface {
 	 * @param id
 	 * @return
 	 */
-	public BookVO findBookById(Integer id);
+	public BookVO findBookById(Long id);
 }
